+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Remove unexpected sleep in end to end test *Profile page should be able to change user password*
 - Fix Formly editableTags type, Formly API check was failing making the use of editableTags impossible
 - Fix Formly editableTags type when using availableOptions, the values of the tags were displayed as literal representation instead of the names
+- Add the possibility to change the CDN URL when executing end to end tests. You can now create *testConf.json* configuration file as a test equivalent of *conf.json* configuration file
 
 ## DEPENDENCIES